move axios client to @/shared
the axios client contains the CSRF token for posting to the backend. Camino Trekker uses this client+token in addition to Camino Creator, so it makes more sense to put it in a folder clearly labelled `shared`
