+++ resources/camino-creator/app.ts
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import { createPinia } from "pinia";
-import { axiosClient } from "./common/axios";
+import { axiosClient } from "@/shared/axios";
 import "leaflet.locatecontrol/dist/L.Control.Locate.min.js";
 import "leaflet-polylinedecorator/dist/leaflet.polylineDecorator.js";
 import "leaflet-draw/dist/leaflet.draw.js";
